--- chain/jsonrpc/jsonrpc-tests/res/genesis_config.json
@@ -1,89 +1,0 @@
-{
-  "protocol_version": 81,
-  "genesis_time": "1970-01-01T00:00:00.000000000Z",
-  "chain_id": "sample",
-  "genesis_height": 0,
-  "num_block_producer_seats": 50,
-  "num_block_producer_seats_per_shard": [
-    50
-  ],
-  "avg_hidden_validator_seats_per_shard": [
-    0
-  ],
-  "dynamic_resharding": false,
-  "protocol_upgrade_stake_threshold": [
-    4,
-    5
-  ],
-  "epoch_length": 500,
-  "gas_limit": 1000000000000000,
-  "min_gas_price": "100000000",
-  "max_gas_price": "10000000000000000000000",
-  "block_producer_kickout_threshold": 90,
-  "chunk_producer_kickout_threshold": 90,
-  "chunk_validator_only_kickout_threshold": 80,
-  "target_validator_mandates_per_shard": 68,
-  "online_min_threshold": [
-    9,
-    10
-  ],
-  "online_max_threshold": [
-    99,
-    100
-  ],
-  "gas_price_adjustment_rate": [
-    1,
-    100
-  ],
-  "validators": [
-    {
-      "account_id": "test.near",
-      "public_key": "ed25519:9BmAFNRTa5mRRXpSAm6MxSEeqRASDGNh2FuuwZ4gyxTw",
-      "amount": "50000000000000000000000000000000"
-    }
-  ],
-  "transaction_validity_period": 100,
-  "protocol_reward_rate": [
-    1,
-    10
-  ],
-  "max_inflation_rate": [
-    1,
-    20
-  ],
-  "total_supply": "2050000000000000000000000000000000",
-  "num_blocks_per_year": 31536000,
-  "protocol_treasury_account": "test.near",
-  "fishermen_threshold": "10000000000000000000000000",
-  "minimum_stake_divisor": 10,
-  "shard_layout": {
-    "V2": {
-      "boundary_accounts": [],
-      "shard_ids": [
-        0
-      ],
-      "id_to_index_map": {
-        "0": 0
-      },
-      "index_to_id_map": {
-        "0": 0
-      },
-      "shards_split_map": null,
-      "shards_parent_map": null,
-      "version": 0
-    }
-  },
-  "num_chunk_only_producer_seats": 300,
-  "minimum_validators_per_shard": 1,
-  "max_kickout_stake_perc": 100,
-  "minimum_stake_ratio": [
-    1,
-    6250
-  ],
-  "shuffle_shard_assignment_for_chunk_producers": false,
-  "use_production_config": false,
-  "num_chunk_producer_seats": 100,
-  "num_chunk_validator_seats": 300,
-  "chunk_producer_assignment_changes_limit": 5,
-  "records": []
-}
